refactor(token): plugin add addTokenSign method and use plugin addTokenSign init token sign
